Remove player module and change the project name (#5)
