+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "push-it"
+name = "drone-target"
 version = "0.1.0"
 authors = ["Nikita Zemtsov <z-nik@ya.ru>"]
 edition = "2018"
