+++ README.md
@@ -1,2 +1,2 @@
-# push-it
-Demo game for 3D game engine [Dotrix](https://github.com/lowenware/dotrix).
+# Drone Target
+A demo game for 3D game engine [Dotrix](https://github.com/lowenware/dotrix).
